Modifying event display on detail page.
